chore(ts): convert sentry app component files
